requirements.txt
discord.py
aioredis